refactor: extract data hook and lineup component
